requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
cryptography==44.0.0

# HTTP and WebSocket
httpx==0.27.2
websockets==13.1
aiohttp==3.10.10

# Database
supabase==2.9.1
postgrest==0.17.2

# Async utilities
asyncio==3.4.3
aiofiles==24.1.0

# Rate limiting
aiolimiter==1.1.0

# Logging and monitoring
structlog==24.4.0
python-json-logger==2.0.7

# Date/time handling
python-dateutil==2.9.0.post0

# Testing (optional, for development)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.30.0

# Type checking (optional, for development)
mypy==1.11.2

# Web server for health checks
fastapi==0.115.4
uvicorn==0.32.0
